[PATCH] 64bit portability fixes
assorted 64bit issues - extraction of number passed in pointer, use of
%zd for size_t, ssize_t instead of int as return value of ->read(), %p
for pointers instead of casting to int and using %x.

+++ drivers/media/dvb/frontends/stv0299.c
@@ -1267,7 +1267,8 @@ static int uni0299_ioctl (struct dvb_frontend *fe, unsigned int cmd, void *arg)
 					    state->tuner_type);
 
 	case FE_DISHNETWORK_SEND_LEGACY_CMD:
-		return stv0299_send_legacy_dish_cmd (i2c, (u32) arg,
+		return stv0299_send_legacy_dish_cmd (i2c,
+						    (u32)(unsigned long)arg,
 						     state->tuner_type);
 
 	case FE_GET_TUNE_SETTINGS:
